Create BulkDeleteAppointments.vbs
Import to vbaProject.OTM in Outlook and create Macros with it. It will delete all selected appointments. This work great if you go on vaccation in a certain week, just select the messages that you want to delete and it will send a response either with or without a message.